fix file import renames
